# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type TokenErrorResponseCode =
  (typeof TokenErrorResponseCode)[keyof typeof TokenErrorResponseCode];

export const TokenErrorResponseCode = {
  token_invalid: "token_invalid",
  token_expired: "token_expired",
  token_used: "token_used",
} as const;

export interface TokenErrorResponse {
  error: string;
  code: TokenErrorResponseCode;
}

export interface Calendar {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  /** Hex color (e.g. "#7c3aed") used to tint events from this calendar */
  color: string;
  timezone: string;
  /** Display name of the calendar owner / source */
  owner: string;
}

export interface Event {
  id: number;
  calendarId: number;
  title: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  location: string | null;
  startAt: string;
  endAt: string;
  allDay: boolean;
  calendar: Calendar;
}

export interface CalendarBreakdown {
  calendar: Calendar;
  eventCount: number;
  upcomingCount: number;
}

export interface DashboardSummary {
  totalCalendars: number;
  totalEvents: number;
  eventsToday: number;
  eventsThisWeek: number;
  nextEvent: Event | null;
  breakdown: CalendarBreakdown[];
}

export interface TopClub {
  clubId: string;
  clubName: string;
  clubSlug: string;
  memberCount: number;
}

export interface MembershipSummary {
  currentAcademicYear: string;
  totalActiveMembers: number;
  totalDuesCollected: number;
  activeClubCount: number;
  topClubs: TopClub[];
}

export type MemberProgram = (typeof MemberProgram)[keyof typeof MemberProgram];

export const MemberProgram = {
  full_time: "full_time",
  evening: "evening",
} as const;

export interface Member {
  id: string;
  fullName: string;
  email: string;
  program: MemberProgram;
  classYear: number;
}

export interface MembershipRecord {
  clubId: string;
  clubName: string;
  clubSlug: string;
  termYears: number;
  amountPaid: number;
  paidAt: string;
  expiresAt: string;
  isActive: boolean;
  academicYear: string;
}

export interface MemberMembershipsResponse {
  member: Member;
  memberships: MembershipRecord[];
}

export type ClubPricing = { [key: string]: number };

export interface Club {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  calendarId: number | null;
  pricing: ClubPricing;
  isActive: boolean;
}

export interface ProgramBreakdown {
  newThisYear: number;
  returning: number;
  total: number;
}

export interface YearOverYearEntry {
  academicYear: string;
  memberCount: number;
}

export interface ClubSummary {
  club: Club;
  currentAcademicYear: string;
  priorAcademicYear: string;
  fullTime: ProgramBreakdown;
  evening: ProgramBreakdown;
  totalMembers: number;
  amountPaid: number;
  paypalFee: number;
  netDues: number;
  yearOverYear: YearOverYearEntry[];
}

export interface RequestAccessBody {
  email: string;
}

export interface RequestAccessResponse {
  message: string;
  /** Only present in dev when RESEND_API_KEY is not set */
  magicLink?: string;
}

export interface RosterMembership {
  termYears: number;
  amountPaid: number;
  paidAt: string;
  expiresAt: string;
  isActive: boolean;
  academicYear: string;
}

export type RosterMemberProgram =
  (typeof RosterMemberProgram)[keyof typeof RosterMemberProgram];

export const RosterMemberProgram = {
  full_time: "full_time",
  evening: "evening",
} as const;

export interface RosterMember {
  fullName: string;
  email: string;
  program: RosterMemberProgram;
  classYear: number;
  memberships: RosterMembership[];
}

export interface ExpiringMember {
  fullName: string;
  email: string;
  expiresAt: string;
}

export interface LapsedMember {
  fullName: string;
  email: string;
  lastPaidAcademicYear: string;
}

export interface RenewalForecast {
  expiringThisYear: ExpiringMember[];
  lapsedLastYear: LapsedMember[];
}

export interface ClubRoster {
  club: Club;
  currentAcademicYear: string;
  members: RosterMember[];
  renewalForecast: RenewalForecast;
}

export type ListEventsParams = {
  /**
   * ISO 8601 inclusive lower bound
   */
  start: string;
  /**
   * ISO 8601 exclusive upper bound
   */
  end: string;
  /**
   * Comma-separated calendar ids to include. Omit to include all.
   */
  calendarIds?: string;
};

export type ListUpcomingEventsParams = {
  limit?: number;
};

export type SearchMembersParams = {
  q: string;
};

export type GetClubRosterParams = {
  /**
   * Single-use magic link token
   */
  token: string;
};

export type DownloadClubRosterCsvParams = {
  /**
   * Magic link token (already-used tokens accepted)
   */
  token: string;
};