# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clubs/{slug}/summary:
    get:
      operationId: getClubSummary
      tags: [clubs]
      summary: New/returning breakdown, dues totals, and year-over-year for a club
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Club summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubSummary"
        "404":
          description: Club not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clubs/{slug}/request-access:
    post:
      operationId: requestRosterAccess
      tags: [clubs]
      summary: Email a magic link to a registered club lead
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestAccessBody"
      responses:
        "200":
          description: Generic success (response is the same whether or not the email matched)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestAccessResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Club not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clubs/{slug}/roster:
    get:
      operationId: getClubRoster
      tags: [clubs]
      summary: Token-gated full member roster with renewal forecast
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: query
          name: token
          required: true
          description: Single-use magic link token
          schema:
            type: string
      responses:
        "200":
          description: Club roster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubRoster"
        "403":
          description: Token invalid, expired, or already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenErrorResponse"
  /clubs/{slug}/roster.csv:
    get:
      operationId: downloadClubRosterCsv
      tags: [clubs]
      summary: Token-gated CSV export of the full member roster
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: query
          name: token
          required: true
          description: Magic link token (already-used tokens accepted)
          schema:
            type: string
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
        "403":
          description: Token invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    TokenErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          enum: [token_invalid, token_expired, token_used]
      required:
        - error
        - code
...
    Club:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        calendarId:
          type: ["integer", "null"]
        pricing:
          type: object
          additionalProperties:
            type: number
        isActive:
          type: boolean
      required:
        - id
        - name
        - slug
        - description
        - calendarId
        - pricing
        - isActive
    ProgramBreakdown:
      type: object
      properties:
        newThisYear:
          type: integer
        returning:
          type: integer
        total:
          type: integer
      required:
        - newThisYear
        - returning
        - total
    YearOverYearEntry:
      type: object
      properties:
        academicYear:
          type: string
        memberCount:
          type: integer
      required:
        - academicYear
        - memberCount
    ClubSummary:
      type: object
      properties:
        club:
          $ref: "#/components/schemas/Club"
        currentAcademicYear:
          type: string
        priorAcademicYear:
          type: string
        fullTime:
          $ref: "#/components/schemas/ProgramBreakdown"
        evening:
          $ref: "#/components/schemas/ProgramBreakdown"
        totalMembers:
          type: integer
        amountPaid:
          type: number
        paypalFee:
          type: number
        netDues:
          type: number
        yearOverYear:
          type: array
          items:
            $ref: "#/components/schemas/YearOverYearEntry"
      required:
        - club
        - currentAcademicYear
        - priorAcademicYear
        - fullTime
        - evening
        - totalMembers
        - amountPaid
        - paypalFee
        - netDues
        - yearOverYear
    RequestAccessBody:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    RequestAccessResponse:
      type: object
      properties:
        message:
          type: string
        magicLink:
          type: string
          description: Only present in dev when RESEND_API_KEY is not set
      required:
        - message
    RosterMembership:
      type: object
      properties:
        termYears:
          type: integer
        amountPaid:
          type: number
        paidAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        isActive:
          type: boolean
        academicYear:
          type: string
      required:
        - termYears
        - amountPaid
        - paidAt
        - expiresAt
        - isActive
        - academicYear
    RosterMember:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        program:
          type: string
          enum: [full_time, evening]
        classYear:
          type: integer
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/RosterMembership"
      required:
        - fullName
        - email
        - program
        - classYear
        - memberships
    ExpiringMember:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        expiresAt:
          type: string
          format: date-time
      required:
        - fullName
        - email
        - expiresAt
    LapsedMember:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        lastPaidAcademicYear:
          type: string
      required:
        - fullName
        - email
        - lastPaidAcademicYear
    RenewalForecast:
      type: object
      properties:
        expiringThisYear:
          type: array
          items:
            $ref: "#/components/schemas/ExpiringMember"
        lapsedLastYear:
          type: array
          items:
            $ref: "#/components/schemas/LapsedMember"
      required:
        - expiringThisYear
        - lapsedLastYear
    ClubRoster:
      type: object
      properties:
        club:
          $ref: "#/components/schemas/Club"
        currentAcademicYear:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/RosterMember"
        renewalForecast:
          $ref: "#/components/schemas/RenewalForecast"
      required:
        - club
        - currentAcademicYear
        - members
        - renewalForecast