# lib/api-zod/src/generated/types/club.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ClubPricing } from "./clubPricing";

export interface Club {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  calendarId: number | null;
  pricing: ClubPricing;
  isActive: boolean;
}