# lib/api-zod/src/generated/types/clubSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { Club } from "./club";
import type { ProgramBreakdown } from "./programBreakdown";
import type { YearOverYearEntry } from "./yearOverYearEntry";

export interface ClubSummary {
  club: Club;
  currentAcademicYear: string;
  priorAcademicYear: string;
  fullTime: ProgramBreakdown;
  evening: ProgramBreakdown;
  totalMembers: number;
  amountPaid: number;
  paypalFee: number;
  netDues: number;
  yearOverYear: YearOverYearEntry[];
}