# lib/api-zod/src/generated/types/programBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ProgramBreakdown {
  newThisYear: number;
  returning: number;
  total: number;
}