# lib/api-zod/src/generated/types/renewalForecast.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ExpiringMember } from "./expiringMember";
import type { LapsedMember } from "./lapsedMember";

export interface RenewalForecast {
  expiringThisYear: ExpiringMember[];
  lapsedLastYear: LapsedMember[];
}