# lib/api-zod/src/generated/types/rosterMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { RosterMemberProgram } from "./rosterMemberProgram";
import type { RosterMembership } from "./rosterMembership";

export interface RosterMember {
  fullName: string;
  email: string;
  program: RosterMemberProgram;
  classYear: number;
  memberships: RosterMembership[];
}