# lib/api-zod/src/generated/types/tokenErrorResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { TokenErrorResponseCode } from "./tokenErrorResponseCode";

export interface TokenErrorResponse {
  error: string;
  code: TokenErrorResponseCode;
}