# lib/api-zod/src/generated/types/yearOverYearEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface YearOverYearEntry {
  academicYear: string;
  memberCount: number;
}